fix: move fetchUser to settings page
